Add new HUD button assets and update tilemaps
Added new HUD button images for game and inventory states, as well as a story image. Updated tilemaps for overworld rooms 3-7 to include new tilesets (slime_blue, apple 16, shield) and placed new tiles in the maps. This enhances the visual interface and introduces new collectible or interactive elements in the game.